Define interleaved writing as a trait
This removes the "foreign" impl on ChunkBuilder.
